move user constants to container-engine-lib
